Move acknowledgment setting to civicuk
